Line and Area renderers refactoring
